refactor: public 메서드로 설정하였습니다.
