+++ src/main/java/chess/domain/pieces/Bishop.java
@@ -1,4 +1,7 @@
-package domain;
+package chess.domain.pieces;
+
+import chess.domain.board.Location;
+import chess.domain.board.State;
 
 public class Bishop extends Pieces {
 
@@ -7,7 +10,7 @@ public class Bishop extends Pieces {
     }
 
     @Override
-    boolean canMove(final char preRank, final char preFile, final Location moveLocation) {
+    public boolean canMove(final char preRank, final char preFile, final Location moveLocation) {
         char moveRank = moveLocation.getRank();
         char moveFile = moveLocation.getFile();
 
@@ -15,12 +18,12 @@ public class Bishop extends Pieces {
     }
 
     @Override
-    boolean isKnight() {
+    public boolean isKnight() {
         return false;
     }
 
     @Override
-    boolean isPawn() {
+    public boolean isPawn() {
         return false;
     }
 }
