+++ src/main/java/chess/domain/pieces/Pawn.java
@@ -1,6 +1,16 @@
-package domain;
+package chess.domain.pieces;
 
-import static domain.Direction.*;
+import chess.domain.board.Location;
+import chess.domain.board.State;
+
+import static chess.domain.additional.Direction.DOUBLE_DOWN;
+import static chess.domain.additional.Direction.DOUBLE_UP;
+import static chess.domain.additional.Direction.DOWN;
+import static chess.domain.additional.Direction.FIRST_DIAGONAL;
+import static chess.domain.additional.Direction.FOURTH_DIAGONAL;
+import static chess.domain.additional.Direction.SECOND_DIAGONAL;
+import static chess.domain.additional.Direction.THIRD_DIAGONAL;
+import static chess.domain.additional.Direction.UP;
 
 public class Pawn extends Pieces {
 
@@ -21,10 +31,15 @@ public class Pawn extends Pieces {
     }
 
     private boolean checkBlackMove(final char preRank, final char preFile, final Location moveLocation) {
-        if (preFile == INIT_BLACK_PAWN_FILE && moveLocation.isSameLocation((char) (preRank + DOUBLE_DOWN.getRow()), (char) (preFile + DOUBLE_DOWN.getColumn()))) {
-            return true;
+        if (preFile == INIT_BLACK_PAWN_FILE) {
+            return checkBlackFirstMove(preRank, preFile, moveLocation) || checkBlackMoveOneStep(preRank, preFile, moveLocation);
         }
+
         return checkBlackMoveOneStep(preRank, preFile, moveLocation);
+    }
+
+    private boolean checkBlackFirstMove(final char preRank, final char preFile, final Location moveLocation) {
+        return moveLocation.isSameLocation((char) (preRank + DOUBLE_DOWN.getRow()), (char) (preFile + DOUBLE_DOWN.getColumn()));
     }
 
     private boolean checkBlackMoveOneStep(final char preRank, final char preFile, final Location moveLocation) {
@@ -34,10 +49,15 @@ public class Pawn extends Pieces {
     }
 
     private boolean checkWhiteMove(final char preRank, final char preFile, final Location moveLocation) {
-        if (preFile == INIT_WHITE_PAWN_FILE && moveLocation.isSameLocation((char) (preRank + DOUBLE_UP.getRow()), (char) (preFile + DOUBLE_UP.getColumn()))) {
-            return true;
+        if (preFile == INIT_WHITE_PAWN_FILE) {
+            return checkWhiteFirstMove(preRank, preFile, moveLocation) || checkWhiteMoveOneStep(preRank, preFile, moveLocation);
         }
+
         return checkWhiteMoveOneStep(preRank, preFile, moveLocation);
+    }
+
+    private boolean checkWhiteFirstMove(final char preRank, final char preFile, final Location moveLocation) {
+        return moveLocation.isSameLocation((char) (preRank + DOUBLE_UP.getRow()), (char) (preFile + DOUBLE_UP.getColumn()));
     }
 
     private boolean checkWhiteMoveOneStep(final char preRank, final char preFile, final Location moveLocation) {
@@ -47,12 +67,12 @@ public class Pawn extends Pieces {
     }
 
     @Override
-    boolean isKnight() {
+    public boolean isKnight() {
         return false;
     }
 
     @Override
-    boolean isPawn() {
+    public boolean isPawn() {
         return true;
     }
 }
