+++ src/main/java/chess/domain/pieces/Rock.java
@@ -1,4 +1,7 @@
-package domain;
+package chess.domain.pieces;
+
+import chess.domain.board.Location;
+import chess.domain.board.State;
 
 public class Rock extends Pieces {
 
@@ -7,17 +10,17 @@ public class Rock extends Pieces {
     }
 
     @Override
-    boolean canMove(final char preRank, final char preFile, final Location moveLocation) {
+    public boolean canMove(final char preRank, final char preFile, final Location moveLocation) {
         return moveLocation.isSameRank(preRank) || moveLocation.isSameFile(preFile);
     }
 
     @Override
-    boolean isKnight() {
+    public boolean isKnight() {
         return false;
     }
 
     @Override
-    boolean isPawn() {
+    public boolean isPawn() {
         return false;
     }
 }
